Add zwim and Russian executables
The zwim executable will read the user's locale to determine what
language to use.

The Russian executable will always output the help message in Russian.
More importantly, the Russian executable is written in the Cyrillic
alphabet, which means that someone with a Russian layout will not need
to switch the keyboard layout in order to type the executable name.